+++ packages/fiori-app-sub-generator/src/fiori-app-generator/prompting.ts
@@ -195,6 +195,7 @@ export async function promptOdataServiceAnswers(
         serviceId: answers.serviceId,
         edmx: answers.metadata,
         annotations: answers.annotations,
+        valueListMetadata: answers.valueListMetadata,
         version: answers.odataVersion,
         capService: answers.capService,
         source: answers.datasourceType,
@@ -303,9 +304,6 @@ export async function createUI5ApplicationPromptOptions(
         [ui5AppInquirerPromptNames.enableTypeScript]: {
             default: defaultPromptValues[ui5AppInquirerPromptNames.enableTypeScript]
         },
-        [ui5AppInquirerPromptNames.enableCodeAssist]: {
-            hide: true
-        },
         [ui5AppInquirerPromptNames.enableVirtualEndpoints]: {
             hide: service.capService?.capType === 'Java'
         }
@@ -407,6 +405,7 @@ function createOdataServicePromptOptions(options: OdataServiceInquirerOptions): 
         [odataServiceInquirerPromptNames.capProject]: {
             capSearchPaths: options.workspaceFolders ?? [],
             defaultChoice: options.capService?.projectPath,
+            useAutoComplete: getHostEnvironment() === hostEnvironment.cli,
             ...options.promptOptions?.capProject
         },
         [odataServiceInquirerPromptNames.capService]: {
@@ -436,6 +435,10 @@ function createOdataServicePromptOptions(options: OdataServiceInquirerOptions): 
             useAutoComplete: !isYUI,
             includeCloudFoundryAbapEnvChoice: true,
             ...options.promptOptions?.systemSelection
+        },
+        [odataServiceInquirerPromptNames.valueHelpDownload]: {
+            hide: true,
+            ...options.promptOptions?.valueHelpDownload
         }
     };
 }