+++ packages/fiori-freestyle-writer/templates/listdetail/add/webapp/Component.js
@@ -55,7 +55,7 @@ function (UIComponent, Device, models, ListSelector, ErrorHandler) {
         getContentDensityClass : function() {
             if (this._sContentDensityClass === undefined) {
                 // check whether FLP has already set the content density class; do nothing in this case
-                // eslint-disable-next-line fiori-custom/sap-no-proprietary-browser-api, fiori-custom/sap-browser-api-warning
+                // eslint-disable-next-line @sap-ux/fiori-tools/sap-no-proprietary-browser-api, @sap-ux/fiori-tools/sap-browser-api-warning
                 if (document.body.classList.contains("sapUiSizeCozy") || document.body.classList.contains("sapUiSizeCompact")) {
                     this._sContentDensityClass = "";
                 } else if (!Device.support.touch) { // apply "compact" mode if touch is not supported