+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/system-selection/prompt-helpers.test.ts
@@ -15,14 +15,16 @@ const backendSystemBasic: BackendSystem = {
     url: 'http://abap.on.prem:1234',
     username: 'user1',
     password: 'password1',
-    systemType: 'OnPrem'
+    systemType: 'OnPrem',
+    connectionType: 'abap_catalog'
 };
 
 const backendSystemReentrance: BackendSystem = {
     name: 'http://s4hc:1234',
     url: 'http:/s4hc:1234',
     authenticationType: 'reentranceTicket',
-    systemType: 'S4HC'
+    systemType: 'AbapCloud',
+    connectionType: 'abap_catalog'
 };
 
 const backendSystems: BackendSystem[] = [backendSystemBasic, backendSystemReentrance];
@@ -83,6 +85,24 @@ describe('Test system selection prompt helpers', () => {
                         type: 'newSystemChoice'
                     }
                 },
+                {
+                    name: backendSystemBasic.name,
+                    value: {
+                        system: backendSystemBasic,
+                        type: 'backendSystem'
+                    }
+                },
+                {
+                    name: `${backendSystemReentrance.name} (ABAP Cloud)`,
+                    value: {
+                        system: backendSystemReentrance,
+                        type: 'backendSystem'
+                    }
+                }
+            ]);
+
+            // hideNewSystem set to true
+            expect(await createSystemChoices(undefined, false, true)).toEqual([
                 {
                     name: backendSystemBasic.name,
                     value: {