+++ packages/project-access/src/index.ts
@@ -1,5 +1,13 @@
-export { FileName, DirName, FioriToolsSettings, MinCdsPluginUi5Version, MinCdsVersion } from './constants';
+export {
+    FileName,
+    DirName,
+    FioriToolsSettings,
+    MinCdsPluginUi5Version,
+    MinCdsVersion,
+    fioriToolsDirectory
+} from './constants';
 export { getFilePaths } from './file';
+export { normalizePath } from './path';
 export {
     addPackageDevDependency,
     clearCdsModuleCache,
@@ -39,6 +47,7 @@ export {
     getMockDataPath,
     getNodeModulesPath,
     getPathMappings,
+    type PathMappings,
     getProject,
     getProjectType,
     getWebappPath,
@@ -55,7 +64,9 @@ export {
     getWorkspaceInfo,
     hasMinCdsVersion,
     checkCdsUi5PluginEnabled,
-    readFlexChanges
+    readFlexChanges,
+    processServices,
+    getMainService
 } from './project';
 export { execNpmCommand } from './command/npm-command';
 export * from './types';